Move variable length match tests into own file
